requirements.txt
coreapi==2.0.9
Django==1.10.2
django-braces==1.10.0
django-filter==0.15.3
django-guardian==1.4.6
django-oauth-toolkit==0.10.0
djangorestframework==3.5.1
djangorestframework-oauth==1.1.0
httpie==0.9.6
itypes==1.1.0
Markdown==2.6.7
oauthlib==1.0.3
Pillow==3.4.2
psycopg2==2.6.2
Pygments==2.1.3
requests==2.11.1
six==1.10.0
uritemplate==3.0.0
